+++ tests/unit/exportplan/test_helpers.py
@@ -167,6 +167,37 @@ def test_update_export_plan(mock_exportplan_update):
     )
 
 
+@mock.patch.object(helpers, 'get_exportplan')
+def get_export_plan_or_create_existing(mock_get_exportplan, user):
+    mock_get_exportplan.return_value = create_response(status_code=200, json_body={'export_plan'})
+
+    export_plan = helpers.get_export_plan_or_create(user)
+
+    assert mock_get_exportplan.call_count == 1
+    assert mock_get_exportplan.call_args == mock.call(sso_session_id=123)
+    assert export_plan == {'export_plan'}
+
+
+@mock.patch.object(helpers, 'get_exportplan')
+@mock.patch.object(helpers, 'get_rules_and_regulations')
+@mock.patch.object(helpers, 'create_export_plan')
+def get_export_plan_or_create_created(
+        mock_get_exportplan, mock_get_rules_and_regulations, mock_create_export_plan, user
+):
+    mock_get_exportplan.return_value = None
+    mock_get_rules_and_regulations.return_value = {
+        'country': 'UK', 'commodity_code': '123', 'rules_regulations': 'abc'
+    }
+    mock_create_export_plan.return_value = {'export_plan_created'}
+
+    export_plan = helpers.get_export_plan_or_create(user)
+
+    assert mock_get_exportplan.call_count == 1
+    assert mock_get_exportplan.call_args == mock.call(sso_session_id=123)
+
+    assert export_plan == {'export_plan_created'}
+
+
 @mock.patch.object(api_client.personalisation, 'recommended_countries_by_sector')
 def test_get_recommended_countries(mock_recommended_countries):
     recommended_countries = [{'country': 'japan'}, {'country': 'south korea'}]
@@ -184,3 +215,18 @@ def test_get_recommended_countries_no_return(mock_recommended_countries):
     countries = helpers.get_recommended_countries(sso_session_id=123, sectors=['Automotive'])
 
     assert countries == []
+
+
+def test_serialize_exportplan_data(user):
+    rules_regulations_data = {
+        'country': 'UK', 'commodity_code': '123'
+    }
+
+    exportplan_data = helpers.serialize_exportplan_data(rules_regulations_data, user)
+
+    assert exportplan_data == {
+        'export_countries': ['UK'],
+        'export_commodity_codes': ['123'],
+        'rules_regulations': {'country': 'UK', 'commodity_code': '123'},
+        'target_markets': [{'country': 'UK'}]
+    }