+++ CHANGELOG.md
@@ -62,4 +62,6 @@
 - no ticket - rm redundant CircleCI step setup_remote_docker
 - MVP-388 - add commdity name
 - MVP-404 - make all tests use PostgreSQL
+- MVP-405 - create initial export plan
+